requirements.txt
flask~=1.1.2
werkzeug~=1.0.1
gunicorn~=20.1.0
MarkupSafe~=2.0.1
cairosvg~=2.5.2
numpy
scikit-image~=0.18.1
svgpathtools~=1.4.1
pillow
setuptools~=52.0.0
pandas~=1.2.1
scikit-learn~=0.24.1
torchvision~=0.12.0
scipy
ipython~=7.20.0
networkx~=2.5
pickle5~=0.0.11
werkzeug~=1.0.1
entmoot~=0.2.2
requests~=2.20.0
pickle5~=0.0.11
svgpathtools~=1.4.1
pandas~=1.2.4
torch~=1.11.0
tqdm~=4.60.0
moviepy~=1.0.3
matplotlib~=3.4.2
scikit-image~=0.18.1
gurobipy~=9.5.0
tensorboardX~=2.2
jupyter~=1.0.0
seaborn~=0.11.1
